Rewrite t-test to compute probability instead of returning difference
The previous API would determine the difference in means at a given
confidence level.  This limited us to the predefined values in the table
(which after testing turns out were wrong anyway.)

The new API determines the probability that the difference is means is
due purely to chance, allowing the user to select the confidence level
and act accordingly.  We're no longer limited to the set of precomputed
ranges.

The new code fails to approximate non-integer degrees of freedom, so we
just truncate in that case.  This is an ok approximation.